refactor: Bank와 BankManager의 역할분리 및 n명의매니저가대응하도록수정
